Minor changes to improve readability of history's queue folder
