+++ service/history/queue/processing_queue_collection.go
@@ -24,16 +24,16 @@ import (
 	"fmt"
 	"sort"
 
+	"github.com/uber/cadence/common/log"
+	"github.com/uber/cadence/common/metrics"
 	"github.com/uber/cadence/service/history/task"
 )
 
-type (
-	processingQueueCollection struct {
-		level       int
-		queues      []ProcessingQueue
-		activeQueue ProcessingQueue
-	}
-)
+type processingQueueCollection struct {
+	level       int
+	queues      []ProcessingQueue
+	activeQueue ProcessingQueue
+}
 
 // NewProcessingQueueCollection creates a new collection for non-overlapping queues
 func NewProcessingQueueCollection(level int, queues []ProcessingQueue) ProcessingQueueCollection {
@@ -43,8 +43,31 @@ func NewProcessingQueueCollection(level int, queues []ProcessingQueue) Processin
 		queues: queues,
 	}
 	queueCollection.resetActiveQueue()
-
 	return queueCollection
+}
+
+func newProcessingQueueCollections(
+	processingQueueStates []ProcessingQueueState,
+	logger log.Logger,
+	metricsClient metrics.Client,
+) []ProcessingQueueCollection {
+	processingQueuesMap := make(map[int][]ProcessingQueue) // level -> state
+	for _, queueState := range processingQueueStates {
+		processingQueuesMap[queueState.Level()] = append(processingQueuesMap[queueState.Level()], NewProcessingQueue(
+			queueState,
+			logger,
+			metricsClient,
+		))
+	}
+	processingQueueCollections := make([]ProcessingQueueCollection, 0, len(processingQueuesMap))
+	for level, queues := range processingQueuesMap {
+		processingQueueCollections = append(processingQueueCollections, NewProcessingQueueCollection(
+			level,
+			queues,
+		))
+	}
+
+	return processingQueueCollections
 }
 
 func (c *processingQueueCollection) Level() int {
@@ -113,9 +136,7 @@ func (c *processingQueueCollection) UpdateAckLevels() (task.Key, int) {
 	return minAckLevel, totalPendingTasks
 }
 
-func (c *processingQueueCollection) Split(
-	policy ProcessingQueueSplitPolicy,
-) []ProcessingQueue {
+func (c *processingQueueCollection) Split(policy ProcessingQueueSplitPolicy) []ProcessingQueue {
 	if len(c.queues) == 0 {
 		return nil
 	}
@@ -142,9 +163,7 @@ func (c *processingQueueCollection) Split(
 	return nextLevelQueues
 }
 
-func (c *processingQueueCollection) Merge(
-	incomingQueues []ProcessingQueue,
-) {
+func (c *processingQueueCollection) Merge(incomingQueues []ProcessingQueue) {
 	sortProcessingQueue(incomingQueues)
 
 	newQueues := make([]ProcessingQueue, 0, len(c.queues)+len(incomingQueues))
