+++ service/history/queue/timer_queue_processor.go
@@ -27,8 +27,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/pborman/uuid"
-
 	"github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
@@ -46,30 +44,28 @@ import (
 	"github.com/uber/cadence/service/worker/archiver"
 )
 
-type (
-	timerQueueProcessor struct {
-		shard         shard.Context
-		historyEngine engine.Engine
-		taskProcessor task.Processor
+type timerQueueProcessor struct {
+	shard         shard.Context
+	historyEngine engine.Engine
+	taskProcessor task.Processor
 
-		config             *config.Config
-		currentClusterName string
+	config             *config.Config
+	currentClusterName string
 
-		metricsClient metrics.Client
-		logger        log.Logger
+	metricsClient metrics.Client
+	logger        log.Logger
 
-		status       int32
-		shutdownChan chan struct{}
-		shutdownWG   sync.WaitGroup
+	status       int32
+	shutdownChan chan struct{}
+	shutdownWG   sync.WaitGroup
 
-		ackLevel               time.Time
-		taskAllocator          TaskAllocator
-		activeTaskExecutor     task.Executor
-		activeQueueProcessor   *timerQueueProcessorBase
-		standbyQueueProcessors map[string]*timerQueueProcessorBase
-		standbyQueueTimerGates map[string]RemoteTimerGate
-	}
-)
+	ackLevel               time.Time
+	taskAllocator          TaskAllocator
+	activeTaskExecutor     task.Executor
+	activeQueueProcessor   *timerQueueProcessorBase
+	standbyQueueProcessors map[string]*timerQueueProcessorBase
+	standbyQueueTimerGates map[string]RemoteTimerGate
+}
 
 // NewTimerQueueProcessor creates a new timer QueueProcessor
 func NewTimerQueueProcessor(
@@ -211,9 +207,7 @@ func (t *timerQueueProcessor) NotifyNewTask(clusterName string, info *hcommon.No
 	standbyQueueProcessor.notifyNewTimers(info.Tasks)
 }
 
-func (t *timerQueueProcessor) FailoverDomain(
-	domainIDs map[string]struct{},
-) {
+func (t *timerQueueProcessor) FailoverDomain(domainIDs map[string]struct{}) {
 	// Failover queue is used to scan all inflight tasks, if queue processor is not
 	// started, there's no inflight task and we don't need to create a failover processor.
 	// Also the HandleAction will be blocked if queue processor processing loop is not running.
@@ -248,6 +242,7 @@ func (t *timerQueueProcessor) FailoverDomain(
 			maxReadLevel = queueReadLevel
 		}
 	}
+	// TODO: Below Add call has no effect, understand the underlying intent and fix it.
 	maxReadLevel.Add(1 * time.Millisecond)
 
 	t.logger.Info("Timer Failover Triggered",
@@ -278,11 +273,7 @@ func (t *timerQueueProcessor) FailoverDomain(
 	failoverQueueProcessor.Start()
 }
 
-func (t *timerQueueProcessor) HandleAction(
-	ctx context.Context,
-	clusterName string,
-	action *Action,
-) (*ActionResult, error) {
+func (t *timerQueueProcessor) HandleAction(ctx context.Context, clusterName string, action *Action) (*ActionResult, error) {
 	var resultNotificationCh chan actionResultNotification
 	var added bool
 	if clusterName == t.currentClusterName {
@@ -423,234 +414,6 @@ func (t *timerQueueProcessor) completeTimer() error {
 	t.ackLevel = newAckLevelTimestamp
 
 	return t.shard.UpdateTimerAckLevel(t.ackLevel)
-}
-
-func newTimerQueueActiveProcessor(
-	clusterName string,
-	shard shard.Context,
-	historyEngine engine.Engine,
-	taskProcessor task.Processor,
-	taskAllocator TaskAllocator,
-	taskExecutor task.Executor,
-	logger log.Logger,
-) *timerQueueProcessorBase {
-	config := shard.GetConfig()
-	options := newTimerQueueProcessorOptions(config, true, false)
-
-	logger = logger.WithTags(tag.ClusterName(clusterName))
-
-	taskFilter := func(taskInfo task.Info) (bool, error) {
-		timer, ok := taskInfo.(*persistence.TimerTaskInfo)
-		if !ok {
-			return false, errUnexpectedQueueTask
-		}
-		if notRegistered, err := isDomainNotRegistered(shard, timer.DomainID); notRegistered && err == nil {
-			logger.Info("Domain is not in registered status, skip task in active timer queue.", tag.WorkflowDomainID(timer.DomainID), tag.Value(taskInfo))
-			return false, nil
-		}
-
-		return taskAllocator.VerifyActiveTask(timer.DomainID, timer)
-	}
-
-	updateMaxReadLevel := func() task.Key {
-		return newTimerTaskKey(shard.UpdateTimerMaxReadLevel(clusterName), 0)
-	}
-
-	updateClusterAckLevel := func(ackLevel task.Key) error {
-		return shard.UpdateTimerClusterAckLevel(clusterName, ackLevel.(timerTaskKey).visibilityTimestamp)
-	}
-
-	updateProcessingQueueStates := func(states []ProcessingQueueState) error {
-		pStates := convertToPersistenceTimerProcessingQueueStates(states)
-		return shard.UpdateTimerProcessingQueueStates(clusterName, pStates)
-	}
-
-	queueShutdown := func() error {
-		return nil
-	}
-
-	return newTimerQueueProcessorBase(
-		clusterName,
-		shard,
-		loadTimerProcessingQueueStates(clusterName, shard, options, logger),
-		taskProcessor,
-		NewLocalTimerGate(shard.GetTimeSource()),
-		options,
-		updateMaxReadLevel,
-		updateClusterAckLevel,
-		updateProcessingQueueStates,
-		queueShutdown,
-		taskFilter,
-		taskExecutor,
-		logger,
-		shard.GetMetricsClient(),
-	)
-}
-
-func newTimerQueueStandbyProcessor(
-	clusterName string,
-	shard shard.Context,
-	historyEngine engine.Engine,
-	taskProcessor task.Processor,
-	taskAllocator TaskAllocator,
-	taskExecutor task.Executor,
-	logger log.Logger,
-) (*timerQueueProcessorBase, RemoteTimerGate) {
-	config := shard.GetConfig()
-	options := newTimerQueueProcessorOptions(config, false, false)
-
-	logger = logger.WithTags(tag.ClusterName(clusterName))
-
-	taskFilter := func(taskInfo task.Info) (bool, error) {
-		timer, ok := taskInfo.(*persistence.TimerTaskInfo)
-		if !ok {
-			return false, errUnexpectedQueueTask
-		}
-		if notRegistered, err := isDomainNotRegistered(shard, timer.DomainID); notRegistered && err == nil {
-			logger.Info("Domain is not in registered status, skip task in standby timer queue.", tag.WorkflowDomainID(timer.DomainID), tag.Value(taskInfo))
-			return false, nil
-		}
-		if timer.TaskType == persistence.TaskTypeWorkflowTimeout ||
-			timer.TaskType == persistence.TaskTypeDeleteHistoryEvent {
-			domainEntry, err := shard.GetDomainCache().GetDomainByID(timer.DomainID)
-			if err == nil {
-				if domainEntry.HasReplicationCluster(clusterName) {
-					// guarantee the processing of workflow execution history deletion
-					return true, nil
-				}
-			} else {
-				if _, ok := err.(*types.EntityNotExistsError); !ok {
-					// retry the task if failed to find the domain
-					logger.Warn("Cannot find domain", tag.WorkflowDomainID(timer.DomainID))
-					return false, err
-				}
-				logger.Warn("Cannot find domain, default to not process task.", tag.WorkflowDomainID(timer.DomainID), tag.Value(timer))
-				return false, nil
-			}
-		}
-		return taskAllocator.VerifyStandbyTask(clusterName, timer.DomainID, timer)
-	}
-
-	updateMaxReadLevel := func() task.Key {
-		return newTimerTaskKey(shard.UpdateTimerMaxReadLevel(clusterName), 0)
-	}
-
-	updateClusterAckLevel := func(ackLevel task.Key) error {
-		return shard.UpdateTimerClusterAckLevel(clusterName, ackLevel.(timerTaskKey).visibilityTimestamp)
-	}
-
-	updateProcessingQueueStates := func(states []ProcessingQueueState) error {
-		pStates := convertToPersistenceTimerProcessingQueueStates(states)
-		return shard.UpdateTimerProcessingQueueStates(clusterName, pStates)
-	}
-
-	queueShutdown := func() error {
-		return nil
-	}
-
-	remoteTimerGate := NewRemoteTimerGate()
-	remoteTimerGate.SetCurrentTime(shard.GetCurrentTime(clusterName))
-
-	return newTimerQueueProcessorBase(
-		clusterName,
-		shard,
-		loadTimerProcessingQueueStates(clusterName, shard, options, logger),
-		taskProcessor,
-		remoteTimerGate,
-		options,
-		updateMaxReadLevel,
-		updateClusterAckLevel,
-		updateProcessingQueueStates,
-		queueShutdown,
-		taskFilter,
-		taskExecutor,
-		logger,
-		shard.GetMetricsClient(),
-	), remoteTimerGate
-}
-
-func newTimerQueueFailoverProcessor(
-	standbyClusterName string,
-	shardContext shard.Context,
-	historyEngine engine.Engine,
-	taskProcessor task.Processor,
-	taskAllocator TaskAllocator,
-	taskExecutor task.Executor,
-	logger log.Logger,
-	minLevel, maxLevel time.Time,
-	domainIDs map[string]struct{},
-) (updateClusterAckLevelFn, *timerQueueProcessorBase) {
-	config := shardContext.GetConfig()
-	options := newTimerQueueProcessorOptions(config, true, true)
-
-	currentClusterName := shardContext.GetService().GetClusterMetadata().GetCurrentClusterName()
-	failoverStartTime := shardContext.GetTimeSource().Now()
-	failoverUUID := uuid.New()
-	logger = logger.WithTags(
-		tag.ClusterName(currentClusterName),
-		tag.WorkflowDomainIDs(domainIDs),
-		tag.FailoverMsg("from: "+standbyClusterName),
-	)
-
-	taskFilter := func(taskInfo task.Info) (bool, error) {
-		timer, ok := taskInfo.(*persistence.TimerTaskInfo)
-		if !ok {
-			return false, errUnexpectedQueueTask
-		}
-		if notRegistered, err := isDomainNotRegistered(shardContext, timer.DomainID); notRegistered && err == nil {
-			logger.Info("Domain is not in registered status, skip task in failover timer queue.", tag.WorkflowDomainID(timer.DomainID), tag.Value(taskInfo))
-			return false, nil
-		}
-		return taskAllocator.VerifyFailoverActiveTask(domainIDs, timer.DomainID, timer)
-	}
-
-	maxReadLevelTaskKey := newTimerTaskKey(maxLevel, 0)
-	updateMaxReadLevel := func() task.Key {
-		return maxReadLevelTaskKey // this is a const
-	}
-
-	updateClusterAckLevel := func(ackLevel task.Key) error {
-		return shardContext.UpdateTimerFailoverLevel(
-			failoverUUID,
-			shard.TimerFailoverLevel{
-				StartTime:    failoverStartTime,
-				MinLevel:     minLevel,
-				CurrentLevel: ackLevel.(timerTaskKey).visibilityTimestamp,
-				MaxLevel:     maxLevel,
-				DomainIDs:    domainIDs,
-			},
-		)
-	}
-
-	queueShutdown := func() error {
-		return shardContext.DeleteTimerFailoverLevel(failoverUUID)
-	}
-
-	processingQueueStates := []ProcessingQueueState{
-		NewProcessingQueueState(
-			defaultProcessingQueueLevel,
-			newTimerTaskKey(minLevel, 0),
-			maxReadLevelTaskKey,
-			NewDomainFilter(domainIDs, false),
-		),
-	}
-
-	return updateClusterAckLevel, newTimerQueueProcessorBase(
-		currentClusterName, // should use current cluster's time when doing domain failover
-		shardContext,
-		processingQueueStates,
-		taskProcessor,
-		NewLocalTimerGate(shardContext.GetTimeSource()),
-		options,
-		updateMaxReadLevel,
-		updateClusterAckLevel,
-		nil,
-		queueShutdown,
-		taskFilter,
-		taskExecutor,
-		logger,
-		shardContext.GetMetricsClient(),
-	)
 }
 
 func loadTimerProcessingQueueStates(
